% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yice_colors.R
\name{yice_colors}
\alias{yice_colors}
\title{Yice Analytics color palette}
\usage{
yice_colors()
}
\value{
A named character vector of hexadecimal colors.
}
\description{
Returns the core Yice Analytics color palette.
}
