% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yice-dependencies.R
\name{yice_dependencies}
\alias{yice_dependencies}
\title{Load Yice Analytics UI dependencies}
\usage{
yice_dependencies()
}
\value{
An HTML dependency object.
}
\description{
Adds the CSS dependencies used by yayice UI components.
}
