% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yice-info.R
\name{yice_info}
\alias{yice_info}
\title{Show yayice package information}
\usage{
yice_info()
}
\value{
Invisibly returns a list containing package information.
}
\description{
Displays basic information about the currently installed
version of the yayice package.
}
