% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yice-options.R
\name{yice_options}
\alias{yice_options}
\title{Get or set yayice options}
\usage{
yice_options(...)
}
\arguments{
\item{...}{Named yayice options to set.}
}
\value{
If no arguments are supplied, returns the current yayice options.
Otherwise invisibly returns the previous values.
}
\description{
Provides a central configuration system for yayice.
}
