% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yice_colors.R
\name{yice_colors}
\alias{yice_colors}
\title{Yice Analytics colour palette}
\usage{
yice_colors()
}
\value{
A named character vector.
}
\description{
Central colour definitions used throughout the yayice UI system.
}
